practice-app/requirements.txt
Django==3.2.3
django-environ==0.4.0
djangorestframework==3.12.4
drf-yasg==1.20.0
urllib3==1.26.5